sytle: Board 엔티티 관련 이름 Post로 모두 변경
