--- src/main/java/com/example/fieldpasserbe/board/repository/CategoryRepositoryJPA.java
@@ -1,4 +1,0 @@
-package com.example.fieldpasserbe.board.repository;
-
-public interface CategoryRepositoryJPA {
-}
