--- src/main/java/com/example/fieldpasserbe/board/repository/DistrictRepositoryJPA.java
@@ -1,4 +1,0 @@
-package com.example.fieldpasserbe.board.repository;
-
-public interface DistrictRepositoryJPA {
-}
