--- src/main/java/com/example/fieldpasserbe/board/service/BoardService.java
@@ -1,6 +1,0 @@
-package com.example.fieldpasserbe.board.service;
-
-public interface BoardService {
-
-    Long countPostById(int id);
-}
