+++ src/main/java/com/example/fieldpasserbe/post/repository/DistrictRepositoryJPA.java
@@ -1,0 +1,4 @@
+package com.example.fieldpasserbe.post.repository;
+
+public interface DistrictRepositoryJPA {
+}
