+++ src/main/java/com/example/fieldpasserbe/post/repository/StadiumRepositoryJPA.java
@@ -1,0 +1,4 @@
+package com.example.fieldpasserbe.post.repository;
+
+public interface StadiumRepositoryJPA {
+}
