+++ src/main/java/com/example/fieldpasserbe/post/service/PostService.java
@@ -1,0 +1,6 @@
+package com.example.fieldpasserbe.post.service;
+
+public interface PostService {
+
+    Long countPostById(int id);
+}
